Refactor rename skip check variable
It's not that the index *was* skipped, it's that it's already been
written, so we don't even need to consider writing to it again. The new
name clarifies this.

The pre-setting of all the values to `false` is unnecessary - if the
key isn't already in the map, golang will return the zero value, which
is `false` in this case.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>