Remove kubeClient global
This was confusingly named, and it was a pointer to an interface, which
is extra confusing. The clientSet (another global) actually implements
that interface, so it is possible to use that and may be a bit cleaner.

It reduces the number of arguments to Initialize, and made it more
obvious that we could re-use the clientset in main.go

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>