Fix bug where parent policy event was not sent
The `parentStatusUpdateNeeded` variable is (usually) carefully guarded
to ensure it only ever moves from false->true, and never back to false.
The moved section of code was not obeying that, and it was being run for
every object template. The result was that one object template could
mark the status as needing an update, but then be overriden by a later
run through the loop.

This section can just be outside of the loop, improving performance and
removing the bug. The additional variable to guard the update to
`parentSatusUpdateNeeded` might not be strictly necessary, but this bug
was too annoying to find to not make every effort to prevent it in the
future.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>